Corriger le router pour afficher galerie et ou un post en fonction de son ID
