+++ src/view/index.twig
@@ -18,13 +18,13 @@
                 <div id="card-container" class="flex-column">
                     <h1 class="profil-name">Lauryann MAGAR</h1>
                     <p>Développeuse PHP / Symfony</p>
-                    <a href="http://localhost:3000/public/blog" class="button-blog">Blog</a>
+                    <a href="http://localhost:3000/blog" class="button-blog">Blog</a>
                 </div>
                 <div class="flex-between footer-profil">
                     <a href="https://github.com/Orcryx" alt="lien vers Github"><i class="fa-brands fa-github"></i></a>
                     <a href="https://www.linkedin.com/in/lmagar/" alt="lien vers Linkedin"><i class="fa-brands fa-linkedin"></i></a>
                     <a href="public/asset/pdf.pdf" alt="lien de téléchargement du CV" download><i class="fa-solid fa-file"></i></a>
-                   <a href="http://localhost:3000/public/contact"><i class="fa-solid fa-message" alt="lien vers le formulaire de contact"></i></a>
+                   <a href="http://localhost:3000/contact"><i class="fa-solid fa-message" alt="lien vers le formulaire de contact"></i></a>
                 </div>
             </div>
         </div>
